Add code and properties to change message popup
Added a code that detects whether the input is valid or not: if invalid, it will show the message popup in the normal color; if valid, it will change all the colors and the path icon. To change the path icon, it's better to just add another path tag and put the svg there, and hide it with display:none until necessary. Also added some ids to the new tags.